+++ src/base/core/src/geom/dir/nav/mod.rs
@@ -1,0 +1,18 @@
+// devela_base_core::geom::dir::nav
+//
+#![doc = crate::_DOC_GEOM_DIR_NAV!()] // private
+#![doc = crate::_doc!(modules: crate::geom::dir; nav)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod face; // DirFace
+
+crate::structural_mods! { // _mods
+    _mods {
+        // #[doc(inline)]
+        // pub use super::{
+        //     face::*,
+        // };
+    }
+}