+++ src/base/core/src/geom/fig/curve/mod.rs
@@ -1,0 +1,14 @@
+// devela_base_core::geom::fig::curve
+//
+#![doc = crate::_DOC_GEOM_FIG_CURVE!()] // public
+#![doc = crate::_doc!(modules: crate::geom::fig; curve)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        // };
+    }
+}