+++ src/base/core/src/geom/space/grid/mod.rs
@@ -1,0 +1,26 @@
+// devela_base_core::geom::space::grid
+//
+#![doc = crate::_DOC_GEOM_SPACE_GRID!()] // public
+#![doc = crate::_doc!(modules: crate::geom::space; grid)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//!
+//! > declarative rules that constrain alignment to grid features
+
+// mod guide; // Reference guides defining alignment and structure in a grid.
+// mod rule; // Snapping rules for aligning positions to grid features.
+// mod schema; // Grid schemas defining rows, columns, modules, and proportions.
+// mod unit; // Grid units and modular measures.
+// mod zone; // Named grid regions and spatial zones.
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     guide::*,
+        //     rule::*,
+        //     schema::*,
+        //     unit::*,
+        //     zone::*,
+        // };
+    }
+}