+++ src/base/core/src/geom/space/motion/mod.rs
@@ -1,0 +1,15 @@
+// devela_base_core::geom::motion
+//
+#![doc = crate::_DOC_GEOM_MOTION!()] // public
+#![doc = crate::_doc!(modules: crate::geom::space; motion)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        // };
+    }
+}