+++ src/base/core/src/geom/space/part/mod.rs
@@ -1,0 +1,21 @@
+// devela_base_core::geom::part
+//
+#![doc = crate::_DOC_GEOM_PART!()] // public
+#![doc = crate::_doc!(modules: crate::geom::space; part)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod cell;
+// mod region;
+// mod test;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     cell::_all::*,
+        //     region::_all::*,
+        //     tess::_all::*,
+        // };
+    }
+}