+++ src/base/core/src/geom/space/topol/mod.rs
@@ -1,0 +1,14 @@
+// devela_base_core::geom::topol
+//
+#![doc = crate::_DOC_GEOM_TOPOL!()] // public
+#![doc = crate::_doc!(modules: crate::geom::space; topol)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        // };
+    }
+}