+++ src/base/core/src/lang/prog/ffi/mod.rs
@@ -1,6 +1,6 @@
 // devela_base_core::lang::prog::ffi
 //
-#![doc = crate::_DOC_LANG_PROG_FFI!()]
+#![doc = crate::_DOC_LANG_PROG_FFI!()] // public
 #![doc = crate::_doc!(modules: crate::lang::prog; ffi: c)] // glsl, js
 #![doc = crate::_doc!(flat:"lang")]
 #![doc = crate::_doc!(extends: ffi)]