+++ src/base/core/src/work/process/mod.rs
@@ -1,5 +1,6 @@
 // devela_base_core::work::process
 //
+#![doc = crate::_doc!(wip)]
 #![doc = crate::_DOC_WORK_PROCESS!()] // public
 #![doc = crate::_doc!(modules: crate::work; process)]
 #![doc = crate::_doc!(flat:"work")]