+++ src/geom/fig/mod.rs
@@ -1,0 +1,23 @@
+// devela::geom::fig
+//
+#![doc = crate::_DOC_GEOM_FIG!()] // public
+#![doc = crate::_doc!(modules: crate::geom; fig)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod curve; // Curved geometric primitives and parametric paths.
+// mod line; // Linear geometric primitives such as segments and rays.
+// mod poly; // Polygonal and polyhedral geometric figures.
+// mod set; // Sets of geometric figures.
+
+crate::structural_mods! { // _mods, _reexports
+    _mods {
+        // pub use super::{
+        // };
+    }
+    _reexports {
+        // use devela_base_core::geom::fig::{
+        // };
+    }
+}