+++ src/geom/fig/poly/mod.rs
@@ -1,0 +1,25 @@
+// devela_base::geom::fig::poly
+//
+#![doc = crate::_DOC_GEOM_FIG_POLY!()] // public
+#![doc = crate::_doc!(modules: crate::geom::fig; poly)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod gon; // polygonal figures (2D)
+// mod hedra; // polyhedral figures (3D)
+// mod rect; // orthogonally constrained gons
+// mod star; // radially constrained gons
+
+crate::structural_mods! { // _mods, _reexports
+    _mods {
+        // pub use super::{
+        //     gon::_all::*,
+        //     hedra::_all::*,
+        //     rect::_all::*,
+        //     star::_all::*,
+        // };
+    }
+    _reexports {
+    }
+}