+++ src/geom/fig/set/mod.rs
@@ -1,0 +1,21 @@
+// devela::geom::fig::set
+//
+#![doc = crate::_DOC_GEOM_FIG_SET!()] // public
+#![doc = crate::_doc!(modules: crate::geom::fig; set)]
+#![doc = crate::_doc!(flat:"geom")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod hull; // convex / concave hulls
+// mod points; //
+
+crate::structural_mods! { // _mods, _reexports
+    _mods {
+        // pub use super::{
+        //     hull::_all::*,
+        //     points::_all::*,
+        // };
+    }
+    _reexports {
+    }
+}